Refactor get_db dependency into app.core.dependencies
